refactor: DRY asset methods
